requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.0
pymupdf>=1.23.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
